Predictions
Ideally I want to generate a response to any sentence the user types in. To do this it needs to remember that this model does not take string input, it takes a bag of words. I also need to realize that the model does not spit out sentences, it generates a list of probabilities for all of classes. This makes the process to generate a response look like the following:
– Get some input from the user
– Convert it to a bag of words
– Get a prediction from the model
– Find the most probable class
– Pick a response from that class